[Silver II] Title: 랜선 자르기, Time: 4 ms, Memory: 2288 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.cc
@@ -26,7 +26,7 @@ int main() {
 
 	while (st <= en) {
 		ll mid = (st + en) / 2;
-		int cnt = 0;
+		ll cnt = 0;
 		for (int i = 0; i < K; i++) {
 			cnt += V[i] / mid;
 		}
